Do not create hash entries during validate.
The new hash ref check in Data::Validate::Struct version 0.11
unconditionally accesses the fields in the hash to be validated.
If the key does not exists, the config tree given to validate()
changes when reading the reference type of the non nonexistent
value.  Only do the type check if the subtree exists.  This restores
old behavior of 0.10 and tests of my OSPF::LSDB module pass again.